Moved map application to helper
